+++ src/main/java/com/delivery/springfood/infrastructure/repository/RestaurantRepositoryImpl.java
@@ -3,6 +3,7 @@ package com.delivery.springfood.infrastructure.repository;
 import com.delivery.springfood.domain.model.Restaurant;
 import com.delivery.springfood.domain.repository.RestaurantRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -34,10 +35,12 @@ public class RestaurantRepositoryImpl implements RestaurantRepository {
     }
 
     @Override
-    @Transactional
-    public String remove(Restaurant restaurant) {
-        restaurant = search(restaurant.getId());
+    public void remove(Long id) {
+        Restaurant restaurant = search(id);
+
+        if (restaurant == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
         entityManager.remove(restaurant);
-        return restaurant.getName();
     }
 }
